fix(docs): keep nb-fs-icon in tree-grid showcase only
