--- src/playground/with-layout/tree-grid/components/fs-icon.component.ts
@@ -1,20 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'nb-fs-icon',
-  template: `
-    <nb-tree-grid-row-toggle [expanded]="expanded" *ngIf="isDir(); else fileIcon">
-    </nb-tree-grid-row-toggle>
-    <ng-template #fileIcon>
-      <span class="nb-compose"></span>
-    </ng-template>
-  `,
-})
-export class FsIconComponent {
-  @Input() kind: string;
-  @Input() expanded: boolean;
-
-  isDir(): boolean {
-    return this.kind === 'dir';
-  }
-}
