+++ src/playground/with-layout/tree-grid/tree-grid-basic.component.ts
@@ -24,8 +24,14 @@ interface FSEntry {
 
           <ng-container [nbTreeGridColumnDef]="customColumn">
             <td nbTreeGridCell *nbTreeGridCellDef="let row">
-              <nb-fs-icon [kind]="row.data.kind" [expanded]="row.expanded"></nb-fs-icon>
+
+              <nb-tree-grid-row-toggle
+                [expanded]="row.expanded"
+                *ngIf="row.data.kind === 'dir'">
+              </nb-tree-grid-row-toggle>
+
               {{row.data.name}}
+
             </td>
           </ng-container>
