+++ src/playground/with-layout/tree-grid/tree-grid-custom-node-structure.component.ts
@@ -22,8 +22,14 @@ interface FSEntry {
           <ng-container [nbTreeGridColumnDef]="customColumn">
             <th nbTreeGridHeaderCell *nbTreeGridHeaderCellDef>{{customColumn}}</th>
             <td nbTreeGridCell *nbTreeGridCellDef="let row">
-              <nb-fs-icon [kind]="row.data.kind" [expanded]="row.expanded"></nb-fs-icon>
+
+              <nb-tree-grid-row-toggle
+                [expanded]="row.expanded"
+                *ngIf="row.data.kind === 'dir'">
+              </nb-tree-grid-row-toggle>
+
               {{row.data[customColumn]}}
+
             </td>
           </ng-container>
