+++ src/playground/with-layout/tree-grid/tree-grid-sortable.component.ts
@@ -30,8 +30,14 @@ interface FSEntry {
             </th>
 
             <td nbTreeGridCell *nbTreeGridCellDef="let row">
-              <nb-fs-icon [kind]="row.data.kind" [expanded]="row.expanded"></nb-fs-icon>
+
+              <nb-tree-grid-row-toggle
+                [expanded]="row.expanded"
+                *ngIf="row.data.kind === 'dir'">
+              </nb-tree-grid-row-toggle>
+
               {{row.data.name}}
+
             </td>
           </ng-container>
